feat: merge UC-04 PostGIS persistence + fix map rendering
- ThuaDat model with PostGIS (geom, centroid, extra fields)
- GIS service: parse dc*.txt/dgn, VN-2000 -> WGS84, store in PostGIS
- REST endpoints: /api/gis/import, /api/gis/parcels, /api/gis/export
- MapView: fix container height (w-full h-full in flex layout)
- MapToolsPanel: collapsible (hidden by default, toggle button)
- Remove debug console.logs from map components
- Add sample dataset (100 parcels, 48 KB)
- Unignore __pycache__ from git tracking
- Reorganize docs: use-cases.md, references.md, README index, architecture.md, feature-ownership.md, security.md, changelog.md
- Resolve merge conflicts with Huy's HoSo/Attachments/DuAn models

+++ backend/app/schemas.py
@@ -56,7 +56,7 @@ class ThuaDatResponse(BaseModel):
     id: int
     tobando: str
     sothua: str
-    dientich: int
+    dientich: float
 
     class Config:
         from_attributes = True
